requirements.txt
flask
flask_cors
pillow
numpy
pandas
opencv-python
matplotlib
seaborn
pexpect
